backend/requirements.txt
Flask
opencv-python
opencv-contrib-python
numpy
Pillow
SQLAlchemy
